Latest version of HQC (2023/04/30) with patches

+++ crypto_kem/hqc-128/clean/code.c
@@ -3,7 +3,6 @@
 #include "reed_muller.h"
 #include "reed_solomon.h"
 #include <stdint.h>
-#include <string.h>
 /**
  * @file code.c
  * @brief Implementation of concatenated code
@@ -19,7 +18,7 @@
  * @param[out] em Pointer to an array that is the tensor code word
  * @param[in] m Pointer to an array that is the message
  */
-void PQCLEAN_HQC128_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
+void PQCLEAN_HQC128_CLEAN_code_encode(uint64_t *em, const uint8_t *m) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
 
     PQCLEAN_HQC128_CLEAN_reed_solomon_encode(tmp, m);
@@ -33,7 +32,7 @@ void PQCLEAN_HQC128_CLEAN_code_encode(uint8_t *em, const uint8_t *m) {
  * @param[out] m Pointer to an array that is the message
  * @param[in] em Pointer to an array that is the code word
  */
-void PQCLEAN_HQC128_CLEAN_code_decode(uint8_t *m, const uint8_t *em) {
+void PQCLEAN_HQC128_CLEAN_code_decode(uint8_t *m, const uint64_t *em) {
     uint8_t tmp[VEC_N1_SIZE_BYTES] = {0};
 
     PQCLEAN_HQC128_CLEAN_reed_muller_decode(tmp, em);
