+++ crypto_kem/hqc-128/clean/code.h
@@ -3,14 +3,13 @@
 
 /**
  * @file code.h
- * Header file of code.c
+ * @brief Header file of code.c
  */
-#include "parameters.h"
-#include <stddef.h>
+
 #include <stdint.h>
 
-void PQCLEAN_HQC128_CLEAN_code_encode(uint8_t *em, const uint8_t *message);
+void PQCLEAN_HQC128_CLEAN_code_encode(uint64_t *em, const uint8_t *message);
 
-void PQCLEAN_HQC128_CLEAN_code_decode(uint8_t *m, const uint8_t *em);
+void PQCLEAN_HQC128_CLEAN_code_decode(uint8_t *m, const uint64_t *em);
 
 #endif
