+++ crypto_kem/hqc-256/clean/reed_muller.h
@@ -3,14 +3,13 @@
 
 /**
  * @file reed_muller.h
- * Header file of reed_muller.c
+ * @brief Header file of reed_muller.c
  */
-#include "parameters.h"
-#include <stddef.h>
+
 #include <stdint.h>
 
-void PQCLEAN_HQC256_CLEAN_reed_muller_encode(uint8_t *cdw, const uint8_t *msg);
+void PQCLEAN_HQC256_CLEAN_reed_muller_encode(uint64_t *cdw, const uint8_t *msg);
 
-void PQCLEAN_HQC256_CLEAN_reed_muller_decode(uint8_t *msg, const uint8_t *cdw);
+void PQCLEAN_HQC256_CLEAN_reed_muller_decode(uint8_t *msg, const uint64_t *cdw);
 
 #endif
